backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy==2.0.27
aiosqlite==0.20.0
alembic==1.13.1
python-multipart==0.0.9
pydantic==2.6.1
pydantic-settings==2.1.0
PyPDF2==3.0.1
python-pptx==0.6.23
sentence-transformers==2.3.1
faiss-cpu==1.7.4
scikit-learn==1.4.0
httpx==0.27.0
python-jose==3.3.0
aiofiles==23.2.1
openai-whisper==20231117
imageio-ffmpeg==0.4.9
